KS-1-1 answer for task

+++ src/main/kotlin/lesson1/lesson1_Task1.kt
@@ -1,6 +1,6 @@
 
 fun main(){
-    val numberOfOrders : Int = 75
+    val numberOfOrders : Byte = 75
     val thanks: String ="Спасибо за покупку, ждем Вас вновь"
